+++ src/RcppExports.cpp
@@ -78,6 +78,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// propdiff_betw_matrices
+NumericMatrix propdiff_betw_matrices(const NumericMatrix& x, const NumericMatrix& y);
+RcppExport SEXP _lineup2_propdiff_betw_matrices(SEXP xSEXP, SEXP ySEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type x(xSEXP);
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type y(ySEXP);
+    rcpp_result_gen = Rcpp::wrap(propdiff_betw_matrices(x, y));
+    return rcpp_result_gen;
+END_RCPP
+}
 // fscale
 NumericVector fscale(const NumericVector& x, const NumericVector& y);
 RcppExport SEXP _lineup2_fscale(SEXP xSEXP, SEXP ySEXP) {
@@ -98,6 +110,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_lineup2_corr_betw_matrices_unpaired_all", (DL_FUNC) &_lineup2_corr_betw_matrices_unpaired_all, 2},
     {"_lineup2_rmsd_betw_matrices", (DL_FUNC) &_lineup2_rmsd_betw_matrices, 2},
     {"_lineup2_mad_betw_matrices", (DL_FUNC) &_lineup2_mad_betw_matrices, 2},
+    {"_lineup2_propdiff_betw_matrices", (DL_FUNC) &_lineup2_propdiff_betw_matrices, 2},
     {"_lineup2_fscale", (DL_FUNC) &_lineup2_fscale, 2},
     {NULL, NULL, 0}
 };