+++ tests/testthat/test-dist_betw_arrays.R
@@ -45,4 +45,19 @@ test_that("dist_betw_arrays() works", {
                          paste0("vv", 1:d4[4]))
     expect_error(dist_betw_arrays(x,y4))
 
+    # proportion differences
+    d <- c(5, 2, 8)
+    x <- array(sample(1:3, prod(d), replace=TRUE), dim=d)
+    y <- array(sample(1:3, prod(d), replace=TRUE), dim=d)
+    dimnames(x) <- dimnames(y) <- list(paste0("ind", 1:d[1]),
+                                       paste0("c", 1:d[2]),
+                                       paste0("v", 1:d[3]))
+    X <- cbind(x[,1,], x[,2,])
+    Y <- cbind(y[,1,], y[,2,])
+    colnames(X) <- colnames(Y) <- paste0("col", 1:prod(d[2:3]))
+    expect_equal(dist_betw_arrays(x, y, "propdiff"),
+                 dist_betw_matrices(X, Y, "propdiff"))
+
+
+
 })